+++ lib/TextLink/TextLink.js
@@ -1,15 +1,15 @@
 /**
- * Link
+ * TextLink
  */
 
 import React, { forwardRef } from 'react';
 import PropTypes from 'prop-types';
-import { Link as RouterLink } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import classnames from 'classnames';
-import css from './Link.css';
+import css from './TextLink.css';
 
-const Link = forwardRef(({ className, href, to, children, ...rest }, ref) => {
-  const Element = to ? RouterLink : 'a';
+const TextLink = forwardRef(({ className, href, to, children, ...rest }, ref) => {
+  const Element = to ? Link : 'a';
 
   const props = {
     className: classnames(css.root, className),
@@ -19,19 +19,19 @@ const Link = forwardRef(({ className, href, to, children, ...rest }, ref) => {
   };
 
   return (
-    <Element data-test-link {...props} ref={ref}>
+    <Element data-test-text-link {...props} ref={ref}>
       {children}
     </Element>
   );
 });
 
-Link.propTypes = {
+TextLink.propTypes = {
   children: PropTypes.node,
   className: PropTypes.string,
   href: PropTypes.string,
   to: PropTypes.string,
 };
 
-Link.displayName = 'Link';
+TextLink.displayName = 'TextLink';
 
-export default Link;
+export default TextLink;
