+++ lib/TextLink/stories/BasicUsage.js
@@ -1,28 +1,29 @@
 /**
- * Link: Basic Usage
+ * TextLink: Basic Usage
  */
 
 import React from 'react';
 import { HashRouter } from 'react-router-dom';
 import Icon from '../../Icon';
 import Tooltip from '../../Tooltip';
-import Link from '../Link';
+import TextLink from '../TextLink';
 
 const BasicUsage = () => (
   <HashRouter>
-    Lorem ipsum dolor sit amet, <Link to="#/some-page">consectetur adipiscing elit</Link>. Mauris nec elementum quam.
+    Lorem ipsum dolor sit amet,{' '}
+    <TextLink to="/some-page">consectetur adipiscing elit</TextLink>. Mauris nec elementum quam.
     <br /><br />
-    <Link target="_blank" rel="noopener noreferrer" href="https://folio.org">
+    <TextLink target="_blank" rel="noopener noreferrer" href="https://folio.org">
       <Icon icon="external-link" iconPosition="end">
           This is an external link
       </Icon>
-    </Link>
+    </TextLink>
     <br /><br />
     <div style={{ width: 300 }}>
-      <Link href="#some-hash">
+      <TextLink href="#some-hash">
         This is a multiline link. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris nec elementum quam.
         Aliquam porta sapien dolor, vel gravida risus condimentum et. Nam eget elit purus.
-      </Link>
+      </TextLink>
     </div>
     <br /><br />
     <Tooltip
@@ -30,13 +31,13 @@ const BasicUsage = () => (
       id="link-tooltip"
     >
       {({ ref, ariaIds }) => (
-        <Link
+        <TextLink
           to="/some-page"
           ref={ref}
           aria-labelledby={ariaIds.text}
         >
           With tooltip
-        </Link>
+        </TextLink>
       )}
     </Tooltip>
   </HashRouter>
