+++ lib/TextLink/stories/TextLink.stories.js
@@ -4,6 +4,6 @@ import withReadme from 'storybook-readme/with-readme';
 import readme from '../readme.md';
 import BasicUsage from './BasicUsage';
 
-storiesOf('Link', module)
+storiesOf('TextLink', module)
   .addDecorator(withReadme(readme))
   .add('Basic Usage', () => <BasicUsage />);
