+++ lib/TextLink/tests/TextLink-test.js
@@ -1,5 +1,5 @@
 /**
- * Link -> tests
+ * TextLink -> tests
  */
 
 import React from 'react';
@@ -8,36 +8,36 @@ import { HashRouter } from 'react-router-dom';
 import { expect } from 'chai';
 import { mount } from '../../../tests/helpers';
 
-import LinkInteractor from './interactor';
-import Link from '../Link';
+import TextLinkInteractor from './interactor';
+import TextLink from '../TextLink';
 
-describe('Link', () => {
-  const link = new LinkInteractor();
+describe('TextLink', () => {
+  const textLink = new TextLinkInteractor();
 
   beforeEach(async () => {
     await mount(
-      <Link
+      <TextLink
         id="some-id"
         href="/hello-world"
       >
-        Link
-      </Link>
+        TextLink
+      </TextLink>
     );
   });
 
   it('Should have an ID', () => {
-    expect(link.id).to.equal('some-id');
+    expect(textLink.id).to.equal('some-id');
   });
 
   describe('If rendered with a "href"-prop', () => {
     beforeEach(async () => {
       await mount(
-        <Link href="/hello-world">Link</Link>
+        <TextLink href="/hello-world">TextLink</TextLink>
       );
     });
 
     it('Should render a native anchor tag', () => {
-      expect(link.tagName).to.equal('a');
+      expect(textLink.tagName).to.equal('a');
     });
   });
 
@@ -45,13 +45,13 @@ describe('Link', () => {
     beforeEach(async () => {
       await mount(
         <HashRouter>
-          <Link to="/hello-world">Link</Link>
+          <TextLink to="/hello-world">TextLink</TextLink>
         </HashRouter>
       );
     });
 
     it('Should render a native anchor tag', () => {
-      expect(link.tagName).to.equal('a');
+      expect(textLink.tagName).to.equal('a');
     });
   });
 });
