+++ lib/TextLink/tests/interactor.js
@@ -15,7 +15,7 @@ function tagName(selector) {
 }
 
 export default interactor(class LinkInteractor {
-  static defaultScope = '[data-test-link]';
+  static defaultScope = '[data-test-text-link]';
 
   id = attribute('id');
   tagName = tagName();
